+++ client/src/Contexts/BeerContext.tsx
@@ -2,6 +2,7 @@ import {
   type Dispatch,
   type SetStateAction,
   createContext,
+  useEffect,
   useState,
 } from "react";
 
@@ -10,6 +11,8 @@ const defaultValue: BeerCountType = {
   setBeerCount: () => {},
   favorites: [],
   toggleFavorite: () => {},
+  breweries: [],
+  setBreweries: () => {},
 };
 
 interface BeerCountType {
@@ -17,11 +20,31 @@ interface BeerCountType {
   setBeerCount: Dispatch<SetStateAction<number>>;
   favorites: string[];
   toggleFavorite: (breweryId: string) => void;
+  setBreweries: Dispatch<SetStateAction<Brewery[]>>;
+  breweries: Brewery[];
 }
-
+interface Brewery {
+  id: string;
+  name: string;
+  brewery_type: string;
+  address: string;
+  city: string;
+  state_province: string;
+  postal_code: string;
+  country: string;
+  website_url: string;
+  state: string;
+}
 const BeerContext = createContext<BeerCountType>(defaultValue);
 
 export const BeerProvider = ({ children }: { children: React.ReactNode }) => {
+  const [breweries, setBreweries] = useState<Brewery[]>([]);
+
+  useEffect(() => {
+    fetch("https://api.openbrewerydb.org/v1/breweries?per_page=200")
+      .then((response) => response.json())
+      .then((data) => setBreweries(data));
+  }, []);
   const [beerCount, setBeerCount] = useState<number>(0);
   const [favorites, setFavorites] = useState<string[]>([]);
 
@@ -37,7 +60,14 @@ export const BeerProvider = ({ children }: { children: React.ReactNode }) => {
 
   return (
     <BeerContext.Provider
-      value={{ beerCount, setBeerCount, favorites, toggleFavorite }}
+      value={{
+        beerCount,
+        setBeerCount,
+        favorites,
+        toggleFavorite,
+        breweries,
+        setBreweries,
+      }}
     >
       {children}
     </BeerContext.Provider>
