+++ client/src/pages/MesFavoris.tsx
@@ -1,36 +1,15 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext, useEffect } from "react";
 import BeerContext from "../Contexts/BeerContext";
 
-interface Brewery {
-  id: string;
-  name: string;
-  brewery_type: string;
-  address: string;
-  city: string;
-  state_province: string;
-
-  postal_code: string;
-  country: string;
-  website_url: string;
-  state: string;
-}
-
 const MesFavoris = () => {
-  const { favorites, toggleFavorite } = useContext(BeerContext); // Utiliser le contexte pour récupérer les favoris et la fonction toggleFavorite
-
-  const [breweries, setBreweries] = useState<Brewery[]>([]);
+  const { favorites, toggleFavorite, breweries, setBreweries } =
+    useContext(BeerContext);
   useEffect(() => {
-    fetch("https://api.openbrewerydb.org/v1/breweries?per_page=200")
-      .then((response) => response.json())
-      .then((data) => {
-        // Filtrer les brasseries favorites par leur ID
-        const favoriteBreweries = data.filter((brewery: Brewery) =>
-          favorites.includes(brewery.id),
-        );
-        setBreweries(favoriteBreweries);
-      });
-  }, [favorites]);
-
+    const favoriteBreweries = breweries.filter((brewery) =>
+      favorites.includes(brewery.id),
+    );
+    setBreweries(favoriteBreweries);
+  });
   return (
     <div>
       <h2>Mes brasseries favorites</h2>
